+++ backend/package.json
@@ -21,10 +21,11 @@
     "test:debug": "node --inspect-brk -r tsconfig-paths/register -r ts-node/register node_modules/.bin/jest --runInBand",
     "test:e2e": "jest --config ./test/jest-e2e.json",
     "typeorm": "ts-node ./node_modules/typeorm/cli",
-    "typeorm:run-migrations": "npm run typeorm migration:run -- -d ./typeOrm.config.ts",
-    "typeorm:generate-migration": "npm run typeorm -- -d ./typeOrm.config.ts migration:generate ./src/migrations/%npm_config_name%",
+    "typeorm:run-migrations": "npm run typeorm migration:run -- -d ./src/config/typeOrm.config.ts",
+    "typeorm:generate-migration": "npm run typeorm -- -d ./src/config/typeOrm.config.ts migration:generate ./src/migrations/%npm_config_name%",
     "typeorm:create-migration": "npm run typeorm -- migration:create ./migrations/%npm_config_name%",
-    "typeorm:revert-migration": "npm run typeorm -- -d ./typeOrm.config.ts migration:revert"
+    "typeorm:revert-migration": "npm run typeorm -- -d ./src/config/typeOrm.config.ts migration:revert"
+  
   },
   "dependencies": {
     "@nestjs-modules/mailer": "^1.8.1",
